Refactor SongQueuer interface
songQueuer is now an internal interface to the song_queue package.
SongQueueManager now controls access to a queue. The interface is also
greatly simplified to support only the minimum set of functionality of a
queue. This is allows for easy swapping of the queue implementation
without having to re-implement the parts that are now a part of
SongQueueManager.