Add mpv delay flag and improve error logging in player
When the player host acts as a snapcast server, there is a delay that
puts the video out of sync with the audio that plays over the clients.
This can be mitigated by using mpv's built in audio delay to sync the
video back inline with the client audio.

Also adds opening a pipe to read mpv's stdout so that any errors in
processing the flags to mpv can be printed out.